# backend/main.go
package main

import (
	"log"
	"net/http"
	"os"

	"github.com/go-chi/chi/v5"
	"github.com/go-chi/chi/v5/middleware"
	"github.com/go-chi/cors"
	httpSwagger "github.com/swaggo/http-swagger"
	_ "github.com/Fajarriswandi/dms-app/backend/docs" // docs is generated by Swag CLI
)

// @title           DMS Backend API
// @version         1.0
// @description     Document Management System Backend API
// @termsOfService  http://swagger.io/terms/

// @contact.name   API Support
// @contact.url    http://www.swagger.io/support
// @contact.email  support@swagger.io

// @license.name  Apache 2.0
// @license.url   http://www.apache.org/licenses/LICENSE-2.0.html

// @host      localhost:8080
// @BasePath  /api/v1

// @securityDefinitions.apikey  BearerAuth
// @in                          header
// @name                        Authorization
// @description                 Type "Bearer" followed by a space and JWT token.
func main() {
	r := chi.NewRouter()

	// Middleware
	r.Use(middleware.Logger)
	r.Use(middleware.Recoverer)
	r.Use(middleware.RequestID)
	r.Use(middleware.RealIP)
	r.Use(SecurityHeadersMiddleware) // Security headers

	// CORS with security improvements
	r.Use(cors.Handler(cors.Options{
		AllowedOrigins:   []string{"http://localhost:5173", "http://localhost:3000"},
		AllowedMethods:   []string{"GET", "POST", "PUT", "DELETE", "OPTIONS", "PATCH"},
		AllowedHeaders:   []string{"Accept", "Authorization", "Content-Type", "X-CSRF-Token", "X-Requested-With"},
		ExposedHeaders:   []string{"Link", "X-Total-Count"},
		AllowCredentials: true,
		MaxAge:           300,
	}))

	// Routes
	r.Get("/", indexHandler)
	r.Get("/health", healthHandler)

	// API v1
	r.Route("/api/v1", func(r chi.Router) {
		r.Get("/", apiInfoHandler)

		// Authentication routes (public)
		r.Post("/auth/register", Register)
		r.Post("/auth/login", Login)

		// Protected routes (require JWT)
		r.Group(func(r chi.Router) {
			r.Use(JWTAuthMiddleware)
			r.Get("/auth/profile", GetProfile)
		})

		// Documents routes (protected)
		r.Group(func(r chi.Router) {
			r.Use(JWTAuthMiddleware)
			r.Get("/documents", getDocumentsHandler)
			r.Get("/documents/{id}", getDocumentHandler)
			r.Post("/documents", createDocumentHandler)
			r.Put("/documents/{id}", updateDocumentHandler)
			r.Delete("/documents/{id}", deleteDocumentHandler)
		})
	})

	// Swagger
	r.Get("/swagger/*", httpSwagger.Handler(
		httpSwagger.URL("http://localhost:8080/swagger/doc.json"),
	))

	port := os.Getenv("PORT")
	if port == "" {
		port = "8080"
	}

	log.Printf("Server starting on port :%s", port)
	log.Printf("Swagger UI available at http://localhost:%s/swagger/index.html", port)
	log.Printf("API Base URL: http://localhost:%s/api/v1", port)
	
	if err := http.ListenAndServe(":"+port, r); err != nil {
		log.Fatal(err)
	}
}

// indexHandler handles the root endpoint
// @Summary      Root endpoint
// @Description  Returns API information
// @Tags         General
// @Accept       json
// @Produce      json
// @Success      200  {object}  map[string]string
// @Router       / [get]
func indexHandler(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusOK)
	w.Write([]byte(`{"message": "DMS Backend API", "version": "1.0.0", "swagger": "/swagger/index.html"}`))
}

// healthHandler handles health check
// @Summary      Health check
// @Description  Returns health status
// @Tags         General
// @Accept       json
// @Produce      json
// @Success      200  {object}  map[string]string
// @Router       /health [get]
func healthHandler(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusOK)
	w.Write([]byte(`{"status": "OK", "service": "dms-backend"}`))
}

// apiInfoHandler returns API information
// @Summary      API Information
// @Description  Returns API version and endpoints
// @Tags         General
// @Accept       json
// @Produce      json
// @Success      200  {object}  map[string]interface{}
// @Router       /api/v1 [get]
func apiInfoHandler(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusOK)
	w.Write([]byte(`{
		"api": "DMS Backend API",
		"version": "1.0.0",
		"endpoints": {
			"documents": "/api/v1/documents",
			"swagger": "/swagger/index.html"
		}
	}`))
}

// Document represents a document
type Document struct {
	ID          string `json:"id"`
	Title       string `json:"title"`
	Description string `json:"description"`
	Content     string `json:"content"`
	CreatedAt   string `json:"created_at"`
	UpdatedAt   string `json:"updated_at"`
}

// getDocumentsHandler returns list of documents
// @Summary      Get all documents
// @Description  Returns a list of all documents (requires authentication)
// @Tags         Documents
// @Accept       json
// @Produce      json
// @Security     BearerAuth
// @Success      200  {array}   Document
// @Failure      401  {object}  ErrorResponse
// @Router       /api/v1/documents [get]
func getDocumentsHandler(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusOK)
	w.Write([]byte(`[
		{
			"id": "1",
			"title": "Sample Document 1",
			"description": "This is a sample document",
			"content": "Document content here...",
			"created_at": "2024-01-01T00:00:00Z",
			"updated_at": "2024-01-01T00:00:00Z"
		},
		{
			"id": "2",
			"title": "Sample Document 2",
			"description": "Another sample document",
			"content": "More document content...",
			"created_at": "2024-01-02T00:00:00Z",
			"updated_at": "2024-01-02T00:00:00Z"
		}
	]`))
}

// getDocumentHandler returns a single document
// @Summary      Get document by ID
// @Description  Returns a single document by ID (requires authentication)
// @Tags         Documents
// @Accept       json
// @Produce      json
// @Security     BearerAuth
// @Param        id   path      string  true  "Document ID"
// @Success      200  {object}  Document
// @Failure      401  {object}  ErrorResponse
// @Failure      404  {object}  ErrorResponse
// @Router       /api/v1/documents/{id} [get]
func getDocumentHandler(w http.ResponseWriter, r *http.Request) {
	id := chi.URLParam(r, "id")
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusOK)
	w.Write([]byte(`{
		"id": "` + id + `",
		"title": "Sample Document",
		"description": "This is a sample document",
		"content": "Document content here...",
		"created_at": "2024-01-01T00:00:00Z",
		"updated_at": "2024-01-01T00:00:00Z"
	}`))
}

// createDocumentHandler creates a new document
// @Summary      Create document
// @Description  Creates a new document (requires authentication)
// @Tags         Documents
// @Accept       json
// @Produce      json
// @Security     BearerAuth
// @Param        document  body      Document  true  "Document object"
// @Success      201       {object}  Document
// @Failure      400       {object}  ErrorResponse
// @Failure      401       {object}  ErrorResponse
// @Router       /api/v1/documents [post]
func createDocumentHandler(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusCreated)
	w.Write([]byte(`{
		"id": "3",
		"title": "New Document",
		"description": "A newly created document",
		"content": "New document content...",
		"created_at": "2024-01-03T00:00:00Z",
		"updated_at": "2024-01-03T00:00:00Z"
	}`))
}

// updateDocumentHandler updates a document
// @Summary      Update document
// @Description  Updates an existing document (requires authentication)
// @Tags         Documents
// @Accept       json
// @Produce      json
// @Security     BearerAuth
// @Param        id        path      string    true  "Document ID"
// @Param        document  body      Document  true  "Document object"
// @Success      200       {object}  Document
// @Failure      401       {object}  ErrorResponse
// @Failure      404       {object}  ErrorResponse
// @Router       /api/v1/documents/{id} [put]
func updateDocumentHandler(w http.ResponseWriter, r *http.Request) {
	id := chi.URLParam(r, "id")
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusOK)
	w.Write([]byte(`{
		"id": "` + id + `",
		"title": "Updated Document",
		"description": "This document has been updated",
		"content": "Updated content...",
		"created_at": "2024-01-01T00:00:00Z",
		"updated_at": "2024-01-03T00:00:00Z"
	}`))
}

// deleteDocumentHandler deletes a document
// @Summary      Delete document
// @Description  Deletes a document by ID (requires authentication)
// @Tags         Documents
// @Accept       json
// @Produce      json
// @Security     BearerAuth
// @Param        id   path      string  true  "Document ID"
// @Success      200  {object}  map[string]string
// @Failure      401  {object}  ErrorResponse
// @Failure      404  {object}  ErrorResponse
// @Router       /api/v1/documents/{id} [delete]
func deleteDocumentHandler(w http.ResponseWriter, r *http.Request) {
	id := chi.URLParam(r, "id")
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusOK)
	w.Write([]byte(`{"message": "Document ` + id + ` deleted successfully"}`))
}